Move the tree model to the UML model
It contains a lot of UML specific information, such as ownership,
and how to deal with specific element types.
